+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/terrains/__init__.py
@@ -23,7 +23,7 @@ There are two main components in this package:
 
 from .height_field import *  # noqa: F401, F403
 from .terrain_generator import TerrainGenerator
-from .terrain_generator_cfg import SubTerrainBaseCfg, TerrainGeneratorCfg
+from .terrain_generator_cfg import FlatPatchSamplingCfg, SubTerrainBaseCfg, TerrainGeneratorCfg
 from .terrain_importer import TerrainImporter
 from .terrain_importer_cfg import TerrainImporterCfg
 from .trimesh import *  # noqa: F401, F403