+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/terrains/config/rough.py
@@ -19,7 +19,6 @@ ROUGH_TERRAINS_CFG = TerrainGeneratorCfg(
     horizontal_scale=0.1,
     vertical_scale=0.005,
     slope_threshold=0.75,
-    difficulty_choices=(0.5, 0.75, 0.9),
     use_cache=False,
     sub_terrains={
         "pyramid_stairs": terrain_gen.MeshPyramidStairsTerrainCfg(